Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -48,7 +48,7 @@ def get_average_age_for_season(data, season)
     if season_num == season
     contestants.each do |contestants_hash|
       num_of_contestants += 1
-      total_age = contestants_hash["age"].to_f
+      total_age = contestants_hash["age"].to_i
     end
   end
 average_age = total_age.to_f / num_of_contestants.to_f
